refactor: restore public facade surface and tighten ownership boundaries
